decouple storage init
