+++ config/request.go
@@ -1,0 +1,38 @@
+// Copyright 2022 The Gidari Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+package config
+
+import (
+	"golang.org/x/time/rate"
+)
+
+// Request is the information needed to query the web API for data to transport.
+type Request struct {
+	// Method is the HTTP(s) method used to construct the http request to fetch data for storage.
+	Method string `yaml:"method"`
+
+	// Endpoint is the fragment of the URL that will be used to request data from the API. This value can include
+	// query parameters.
+	Endpoint string `yaml:"endpoint"`
+
+	// Query represent the query params to apply to the URL generated by the request.
+	Query map[string]string
+
+	// Timeseries indicates that the underlying data should be queries as a time series. This means that the
+	Timeseries *Timeseries `yaml:"timeseries"`
+
+	// Table is the name of the table/collection to insert the data fetched from the web API.
+	Table string `yaml:"table"`
+
+	// Truncate before upserting on single request
+	Truncate *bool `yaml:"truncate"`
+
+	// Chunks of requests should share a rate limiter, probably all of them; inheriting the rate limiter from the
+	// root configuration.
+	RateLimiter *rate.Limiter
+}
