+++ config/timeseries.go
@@ -1,0 +1,30 @@
+// Copyright 2022 The Gidari Authors.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+package config
+
+import (
+	"time"
+)
+
+// Timeseries is a struct that contains the information needed to query a web API for Timeseries data.
+type Timeseries struct {
+	StartName string `yaml:"startName"`
+	EndName   string `yaml:"endName"`
+
+	// Period is the size of each chunk in seconds for which we can query the API. Some API will not allow us to
+	// query all data within the start and end range.
+	Period int32 `yaml:"period"`
+
+	// Layout is the time layout for parsing the "Start" and "End" values into "time.Time". The default is assumed
+	// to be RFC3339.
+	Layout *string `yaml:"layout"`
+
+	// Chunks are the time ranges for which we can query the API. These are broken up into pieces for API requests
+	// that only return a limited number of results.
+	Chunks [][2]time.Time
+}
