+++ internal/repository/generic.go
@@ -11,8 +11,7 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/alpstable/gidari/internal/storage"
-	"github.com/alpstable/gidari/proto"
+	proto "github.com/alpstable/gidari-proto"
 )
 
 // ErrFailedToCreateRepository is returned when the repository layer fails to create a new repository.
@@ -26,36 +25,26 @@ func FailedToCreateRepositoryError(err error) error {
 
 // Generic is the interface for the generic service.
 type Generic interface {
-	storage.Storage
-	storage.Transactor
+	proto.Storage
+	proto.Transactor
 
 	Transact(fn func(ctx context.Context, repo Generic) error)
 }
 
 // GenericService is the implementation of the Generic service.
 type GenericService struct {
-	storage.Storage
-	*storage.Txn
+	proto.Storage
+	*proto.Txn
 }
 
 // New returns a new Generic service.
-func New(ctx context.Context, dns string) (*GenericService, error) {
-	stg, err := storage.New(ctx, dns)
-	if err != nil {
-		return nil, fmt.Errorf("failed to construct storage: %w", err)
-	}
-
+func New(ctx context.Context, stg proto.Storage) (*GenericService, error) {
 	return &GenericService{stg, nil}, nil
 }
 
 // NewTx returns a new Generic service with an initialized transaction object that can be used to commit or rollback
 // storage operations made by the repository layer.
-func NewTx(ctx context.Context, dns string) (*GenericService, error) {
-	stg, err := storage.New(ctx, dns)
-	if err != nil {
-		return nil, fmt.Errorf("failed to construct storage: %w", err)
-	}
-
+func NewTx(ctx context.Context, stg proto.Storage) (*GenericService, error) {
 	tx, err := stg.StartTx(ctx)
 	if err != nil {
 		return nil, fmt.Errorf("failed to start transaction: %w", err)
@@ -67,7 +56,7 @@ func NewTx(ctx context.Context, dns string) (*GenericService, error) {
 // Transact is a helper function that wraps a function in a transaction and commits or rolls back the transaction. If
 // svc is not a transaction, the function will be executed without executing.
 func (svc *GenericService) Transact(fn func(ctx context.Context, repo Generic) error) {
-	svc.Txn.Send(func(ctx context.Context, stg storage.Storage) error {
+	svc.Txn.Send(func(ctx context.Context, stg proto.Storage) error {
 		err := fn(ctx, svc)
 		if err != nil {
 			return fmt.Errorf("error executing transaction: %w", err)
