--- internal/storage/queries/pg_garbage_collect.sql
@@ -1,1 +1,0 @@
-VACUUM FULL %s;
