--- internal/storage/queries/pg_truncate_tables.sql
@@ -1,1 +1,0 @@
-TRUNCATE TABLE %s;
