--- internal/transport/testdata/upsert/histam.yml
@@ -1,13 +1,0 @@
-url: https://chroniclingamerica.loc.gov
-connectionStrings:
-  - mongodb://mongo1:27017/histam
-rateLimit:
-  burst: 1
-  period: 1
-requests:
-  -
-    endpoint: /search/titles/results
-    query:
-      terms: oakland
-      format: json
-      page: 5
