--- internal/transport/testdata/upsert/zippopotam.yml
@@ -1,10 +1,0 @@
-url: http://api.zippopotam.us
-connectionStrings:
-  - mongodb://mongo1:27017/zippopotam
-rateLimit:
-  burst: 1
-  period: 1
-requests:
-  -
-    endpoint: /us/98121
-    table: seatle
