--- internal/transport/testdata/upsert/zippopotam_truncate_req.yml
@@ -1,16 +1,0 @@
-url: http://api.zippopotam.us
-connectionStrings:
-  - mongodb://mongo1:27017/zippopotam
-rateLimit:
-  burst: 1
-  period: 1
-requests:
-  -
-    endpoint: /us/98121
-    table: seatle
-    truncate: false
-  -
-    endpoint: /us/94701
-    table: berkeley
-    truncate: true
-
