--- proto/db.proto
@@ -1,70 +1,0 @@
-syntax = "proto3";
-import "google/protobuf/struct.proto";
-
-package proto;
-
-option go_package = ".;proto";
-
-// Create a record in the database. Optionally include an "id" field otherwise it's set automatically.
-message UpsertRequest {
-	string table = 1;
-	int32 dataType = 3;
-	bytes data = 4;
-}
-
-message UpsertResponse {
-	// Number of records upserted
-	int64 upsertedCount = 1;
-
-	// Number of records matched
-	int64 matchedCount = 2;
-}
-
-message Columns {
-	repeated string list = 1;
-}
-
-message ListColumnsResponse {
-	map<string, Columns> colSet = 1;
-}
-
-message PrimaryKeys {
-	repeated string list = 1;
-}
-
-message ListPrimaryKeysResponse {
-	map<string,  PrimaryKeys> PKSet = 3;
-}
-
-message Table {
-	int64 size = 1;
-}
-
-message ListTablesResponse {
-	map<string, Table> tableSet = 1;
-}
-
-// Read data from a table. Lookup can be by ID or via querying any field in the record.
-message ReadRequest {
-	// Optional table name. Defaults to 'default'
-	bytes readerBuilder = 1;
-
-	google.protobuf.Struct required = 2;
-	google.protobuf.Struct options = 3;
-	string table = 4;
-}
-
-message ReadResponse {
-	// JSON encoded records
-	repeated google.protobuf.Struct  records = 1;
-}
-
-message TruncateRequest {
-	// Optional table name. Defaults to 'default'
-	repeated string tables = 1;
-}
-
-message TruncateResponse {
-	// Number of records deleted
-	int32 deletedCount = 1;
-}
